+++ packages/solid/src/main.tsx
@@ -93,6 +93,7 @@ import NewCustomizingDemo from "./components/NewCustomizingDemo";
 import NewMaskInputDemo from "./components/NewMaskInputDemo";
 import NewLinkDemo from "./components/NewLinkDemo";
 import NewThemeDemo from "./components/NewThemeDemo";
+import NewDirectionDemo from "./components/NewDirectionDemo";
 import { DirectionProvider } from "./components/direction/direction";
 import NewColorPickerDemo from "./components/NewColorPickerDemo";
 import NewCarouselDemo from "./components/NewCarouselDemo";
@@ -202,6 +203,7 @@ render(
       <Route path="/mask-input" component={NewMaskInputDemo} />
       <Route path="/link" component={NewLinkDemo} />
       <Route path="/theme" component={NewThemeDemo} />
+      <Route path="/direction" component={NewDirectionDemo} />
       <Route path="/color-picker" component={NewColorPickerDemo} />
       <Route path="/carousel" component={NewCarouselDemo} />
       <Route path="/dynamic-date-range" component={NewDynamicDateRangeDemo} />